Moved code that be used per all class of Modules
