+++ BankApp/DataTests/Comons/XCTestCase+Extension.swift
@@ -1,0 +1,17 @@
+//
+//  XCTestCase+Extension.swift
+//  DataTests
+//
+//  Created by Estaife Lima on 30/05/20.
+//  Copyright © 2020 Estaife Lima. All rights reserved.
+//
+
+import XCTest
+
+public extension XCTestCase {
+    func memoryLeakCheckWith(instance: AnyObject, file: StaticString = #file, line: UInt = #line) {
+        addTeardownBlock { [weak instance] in
+            XCTAssertNil(instance, file: file, line: line)
+        }
+    }
+}
